requirements.txt
beautifulsoup4==4.6.3
environs==4.0.0
Flask==1.0.2
flask_bootstrap
flask_socketio==3.0.2
fpdf==1.7.2
gevent-websocket
image==1.5.24
Jinja2==2.10
Pillow==5.1.0
requests==2.19.1
